collapsed io-extensions and -extensions into arrow-effects-data
