+++ modules/ank/arrow-ank/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'kotlin-kapt'
 
 dependencies {
     compile project(":arrow-core")
-    compile project(":arrow-effects-io-extensions")
+    compile project(":arrow-effects-data")
 
     compile "org.jetbrains.kotlin:kotlin-compiler:$ankKotlinVersion"
     compile "org.jetbrains.kotlin:kotlin-script-util:$ankKotlinVersion"
