+++ modules/docs/arrow-docs/build.gradle
@@ -11,7 +11,7 @@ dependencies {
     compile project(':arrow-ank')
     compile project(':arrow-core-data')
     compile project(':arrow-test')
-    compile project(':arrow-effects-extensions')
+    compile project(':arrow-effects-data')
     compile project(':arrow-effects-rx2-extensions')
     compile project(':arrow-effects-reactor-extensions')
     compile project(':arrow-optics')
