+++ modules/docs/arrow-docs/docs/docs/README.md
@@ -94,8 +94,6 @@ dependencies {
     compile "io.arrow-kt:arrow-mtl-data:$arrow_version" //optional
     compile "io.arrow-kt:arrow-mtl:$arrow_version" //optional
     compile "io.arrow-kt:arrow-effects-data:$arrow_version" //optional
-    compile "io.arrow-kt:arrow-effects-extensions:$arrow_version" //optional
-    compile "io.arrow-kt:arrow-effects-io-extensions:$arrow_version" //optional
     compile "io.arrow-kt:arrow-effects-rx2-data:$arrow_version" //optional
     compile "io.arrow-kt:arrow-effects-rx2-extensions:$arrow_version" //optional
     compile "io.arrow-kt:arrow-effects-reactor-data:$arrow_version" //optional
