+++ modules/docs/arrow-examples/build.gradle
@@ -23,7 +23,7 @@ idea {
 }
 
 dependencies {
-    compile project(':arrow-effects-extensions')
+    compile project(':arrow-effects-data')
     compile project(':arrow-effects-rx2-extensions')
     compile project(':arrow-optics')
     compile project(':arrow-ui')
