+++ modules/effects/arrow-effects-data/src/main/kotlin/arrow/effects/Resource.kt
@@ -29,7 +29,7 @@ inline fun <F, E, A> ResourceOf<F, E, A>.fix(): Resource<F, E, A> =
  * Consider the following use case:
  * ```kotlin:ank:playground
  * import arrow.effects.IO
- * // <fail import>
+ * import arrow.effects.extensions.fx
  *
  * object Consumer
  * object Handle
