+++ modules/effects/arrow-effects-data/src/test/kotlin/arrow/effects/IOTest.kt
@@ -10,7 +10,6 @@ import arrow.core.right
 import arrow.effects.IO.Companion.effect
 import arrow.effects.IO.Companion.just
 import arrow.effects.extensions.fx
-// <fail import>
 import arrow.effects.extensions.io.async.async
 import arrow.effects.extensions.io.concurrent.concurrent
 import arrow.effects.extensions.io.concurrent.parMapN
