--- modules/effects/arrow-effects-extensions/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Arrow-Effects-Extensions
-POM_ARTIFACT_ID=arrow-effects-extensions
-POM_PACKAGING=jar
