--- modules/effects/arrow-effects-io-extensions/build.gradle
@@ -1,17 +1,0 @@
-apply plugin: 'kotlin-kapt'
-
-dependencies {
-    compile project(':arrow-effects-data')
-    compile project(':arrow-effects-extensions')
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
-    compile project(':arrow-annotations')
-    kapt project(':arrow-meta')
-    kaptTest project(':arrow-meta')
-    compileOnly project(':arrow-meta')
-    testCompileOnly project(':arrow-meta')
-    testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
-    testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
-    testCompile project(':arrow-test')
-}
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
