--- modules/effects/arrow-effects-io-extensions/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Arrow-Effects-IO-Extensions
-POM_ARTIFACT_ID=arrow-effects-io-extensions
-POM_PACKAGING=jar
