+++ modules/integrations/arrow-integrations-retrofit-adapter/build.gradle
@@ -15,7 +15,7 @@ dependencies {
     testCompile "com.squareup.retrofit2:converter-gson:$retrofitVersion"
     testCompile 'com.squareup.okhttp3:mockwebserver:3.13.1'
     testCompile project(':arrow-effects-rx2-extensions')
-    testCompile project(':arrow-effects-extensions')
+    testCompile project(':arrow-effects-data')
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
