+++ modules/reflect/arrow-reflect/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'kotlin-kapt'
 dependencies {
     compile 'io.github.classgraph:classgraph:4.8.14'
     compile project(':arrow-core')
-    compile project(':arrow-effects-extensions')
+    compile project(':arrow-effects-data')
     compile project(':arrow-syntax')
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile "org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion"
