+++ modules/streams/arrow-streams/build.gradle
@@ -6,7 +6,7 @@ dependencies {
     testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
     testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
     testCompile project(':arrow-test')
-    testCompile project(':arrow-effects-extensions')
+    testCompile project(':arrow-effects-data')
 
     compile project(':arrow-annotations')
     kapt project(':arrow-meta')
