+++ settings.gradle
@@ -67,8 +67,6 @@ include {
 
         effects {
             _ 'effects-data'
-            _ 'effects-extensions'
-            _ 'effects-io-extensions'
             _ 'effects-rx2-data'
             _ 'effects-rx2-extensions'
             _ 'effects-reactor-data'
